[vfs] E: Split File Domain Types
Move File, OpenOptions, Stat, and DirEntry into dedicated modules while
retaining the existing file facade and public re-exports. This
establishes the one-domain-type-per-module layout without changing VFS
behavior or public paths.
