+++ src/libs/vfs/src/file.rs
@@ -11,6 +11,26 @@
 //!   [`chdir()`], [`cwd()`].
 
 //==================================================================================================
+// Modules
+//==================================================================================================
+
+mod dir_entry;
+mod handle;
+mod open_options;
+mod stat;
+
+//==================================================================================================
+// Re-Exports
+//==================================================================================================
+
+pub use self::{
+    dir_entry::DirEntry,
+    handle::File,
+    open_options::OpenOptions,
+    stat::Stat,
+};
+
+//==================================================================================================
 // Imports
 //==================================================================================================
 
@@ -23,338 +43,6 @@ use ::fat32::{
     Fat32Error,
     FatFile,
 };
-use ::sysapi::unistd::file_seek;
-
-//==================================================================================================
-// OpenOptions
-//==================================================================================================
-
-/// Builder for opening files with specific access options.
-///
-/// Provides a readable, builder-pattern API for specifying file open modes.
-///
-/// # Default Behavior
-///
-/// If you call `open()` without setting any options, it defaults to read-only
-/// mode (equivalent to `.read(true)`).
-///
-/// # Description
-///
-/// ```ignore
-/// use vfs::{OpenOptions, File};
-///
-/// // Open for reading (implicit default)
-/// let file = OpenOptions::new().open("/data/config.txt")?;
-///
-/// // Create a new file for writing
-/// let file = OpenOptions::new()
-///     .write(true)
-///     .create(true)
-///     .open("/data/output.txt")?;
-///
-/// // Create new file, fail if exists (O_CREAT | O_EXCL)
-/// let file = OpenOptions::new()
-///     .write(true)
-///     .create_new(true)
-///     .open("/data/unique.txt")?;
-/// ```
-#[derive(Debug, Clone, Copy, Default)]
-pub struct OpenOptions {
-    read: bool,
-    write: bool,
-    create: bool,
-    create_new: bool,
-    truncate: bool,
-}
-
-impl OpenOptions {
-    /// Creates a new `OpenOptions` with all options set to false.
-    #[must_use]
-    pub const fn new() -> Self {
-        Self {
-            read: false,
-            write: false,
-            create: false,
-            create_new: false,
-            truncate: false,
-        }
-    }
-
-    /// Sets the option for read access.
-    #[must_use]
-    pub const fn read(mut self, read: bool) -> Self {
-        self.read = read;
-        self
-    }
-
-    /// Sets the option for write access.
-    #[must_use]
-    pub const fn write(mut self, write: bool) -> Self {
-        self.write = write;
-        self
-    }
-
-    /// Sets the option to create a new file if it doesn't exist.
-    #[must_use]
-    pub const fn create(mut self, create: bool) -> Self {
-        self.create = create;
-        self
-    }
-
-    /// Sets the option to truncate the file to zero length on open.
-    ///
-    /// Requires `write(true)`.
-    #[must_use]
-    pub const fn truncate(mut self, truncate: bool) -> Self {
-        self.truncate = truncate;
-        self
-    }
-
-    /// Sets the option to create a new file, failing if it already exists.
-    ///
-    /// This is equivalent to `O_CREAT | O_EXCL` in POSIX terms.
-    #[must_use]
-    pub const fn create_new(mut self, create_new: bool) -> Self {
-        self.create_new = create_new;
-        self
-    }
-
-    /// Opens the file at the specified path with the configured options.
-    ///
-    /// If neither `read` nor `write` is set, defaults to `read(true)`.
-    ///
-    /// # Parameters
-    ///
-    /// - `path`: The path to the file to open.
-    ///
-    /// # Returns
-    ///
-    /// A new [`File`] handle, or an error.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::NotInitialized`] if the filesystem hasn't been
-    ///   initialized.
-    /// - [`Fat32Error::NotFound`] if the path doesn't exist and `create` is
-    ///   false.
-    /// - [`Fat32Error::ReadOnly`] if write/create/truncate on a read-only mount.
-    /// - [`Fat32Error::InvalidArgument`] if `truncate` is set without `write`,
-    ///   or if `create_new` is combined with `create` or `truncate`.
-    /// - [`Fat32Error::AlreadyExists`] if `create_new` is set and file exists.
-    pub fn open(self, path: &str) -> Result<File, Fat32Error> {
-        // Validate: truncate requires write.
-        if self.truncate && !self.write {
-            return Err(Fat32Error::InvalidArgument);
-        }
-
-        // Validate: create_new is mutually exclusive with create and truncate.
-        if self.create_new && (self.create || self.truncate) {
-            return Err(Fat32Error::InvalidArgument);
-        }
-
-        // Default to read if neither read nor write specified.
-        let read: bool = if !self.read && !self.write {
-            true
-        } else {
-            self.read
-        };
-
-        open_with_options(path, read, self.write, self.create, self.create_new, self.truncate)
-    }
-}
-
-//==================================================================================================
-// File
-//==================================================================================================
-
-/// An open file handle on a FAT filesystem.
-///
-/// Provides POSIX-like read, write, and seek operations.
-///
-/// # Lifetime
-///
-/// The `'static` lifetime on the inner `FatFile` is safe because the
-/// FAT filesystem backing memory is mapped before guest execution and
-/// lives for the program's entire lifetime.
-///
-/// # Description
-///
-/// ```ignore
-/// use vfs;
-///
-/// let mut file = vfs::open("/data/hello.txt")?;
-/// let mut buf = [0u8; 256];
-/// let n = file.read(&mut buf)?;
-/// ```
-pub struct File {
-    /// The underlying FAT file handle.
-    inner: FatFile<'static>,
-    /// The mount path this file belongs to (for open file tracking).
-    mount_path: String,
-}
-
-impl File {
-    /// Returns true if this file supports writing.
-    #[must_use]
-    pub fn is_writable(&self) -> bool {
-        self.inner.can_write()
-    }
-
-    /// Returns true if this file supports reading.
-    #[must_use]
-    pub fn is_readable(&self) -> bool {
-        self.inner.can_read()
-    }
-
-    /// Reads data from the file.
-    ///
-    /// # Parameters
-    ///
-    /// - `buf`: The buffer to read data into.
-    ///
-    /// # Returns
-    ///
-    /// The number of bytes read, or 0 at EOF.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::PermissionDenied`] if file is not open for reading.
-    /// - [`Fat32Error::IoError`] on read failure.
-    pub fn read(&mut self, buf: &mut [u8]) -> Result<usize, Fat32Error> {
-        self.inner.read(buf)
-    }
-
-    /// Writes data to the file.
-    ///
-    /// # Parameters
-    ///
-    /// - `buf`: The data to write.
-    ///
-    /// # Returns
-    ///
-    /// The number of bytes written.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::ReadOnly`] if file is not open for writing.
-    /// - [`Fat32Error::NoSpace`] if filesystem is full.
-    /// - [`Fat32Error::IoError`] on write failure.
-    pub fn write(&mut self, buf: &[u8]) -> Result<usize, Fat32Error> {
-        self.inner.write(buf)
-    }
-
-    /// Seeks to a position in the file.
-    ///
-    /// # Parameters
-    ///
-    /// - `whence`: Seek mode (`SEEK_SET`, `SEEK_CUR`, or `SEEK_END`).
-    /// - `offset`: Offset in bytes.
-    ///
-    /// # Returns
-    ///
-    /// The new file position.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::InvalidArgument`] if `whence` is invalid or offset is
-    ///   negative for `SEEK_SET`.
-    /// - [`Fat32Error::IoError`] if seeking to an invalid position.
-    pub fn seek(&mut self, whence: i32, offset: i64) -> Result<u64, Fat32Error> {
-        let pos: ::fatfs::SeekFrom = match whence {
-            file_seek::SEEK_SET => {
-                if offset < 0 {
-                    return Err(Fat32Error::InvalidArgument);
-                }
-                ::fatfs::SeekFrom::Start(offset as u64)
-            },
-            file_seek::SEEK_CUR => ::fatfs::SeekFrom::Current(offset),
-            file_seek::SEEK_END => ::fatfs::SeekFrom::End(offset),
-            _ => return Err(Fat32Error::InvalidArgument),
-        };
-        self.inner.seek(pos)
-    }
-
-    /// Flushes any buffered data to the filesystem.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::IoError`] on flush failure.
-    pub fn flush(&mut self) -> Result<(), Fat32Error> {
-        self.inner.flush()
-    }
-
-    /// Gets the file size in bytes.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::IoError`] if seeking fails.
-    pub fn size(&mut self) -> Result<u64, Fat32Error> {
-        self.inner.len()
-    }
-
-    /// Truncates the file at the current position.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::ReadOnly`] if file is not open for writing.
-    /// - [`Fat32Error::IoError`] on truncate failure.
-    pub fn truncate(&mut self) -> Result<(), Fat32Error> {
-        self.inner.truncate()
-    }
-
-    /// Reads the entire file contents into a newly allocated `Vec`.
-    ///
-    /// Seeks to the beginning of the file, then reads until EOF. After
-    /// returning, the file position is at the end of the data that was read.
-    ///
-    /// # Returns
-    ///
-    /// A `Vec<u8>` containing the file contents.
-    ///
-    /// # Errors
-    ///
-    /// - [`Fat32Error::PermissionDenied`] if file is not open for reading.
-    /// - [`Fat32Error::OutOfMemory`] if the file size exceeds addressable memory.
-    /// - [`Fat32Error::IoError`] on read failure.
-    pub fn read_to_vec(&mut self) -> Result<Vec<u8>, Fat32Error> {
-        if !self.inner.can_read() {
-            return Err(Fat32Error::PermissionDenied);
-        }
-
-        let file_size: u64 = self.inner.seek(::fatfs::SeekFrom::End(0))?;
-        self.inner.seek(::fatfs::SeekFrom::Start(0))?;
-
-        let buf_size: usize = usize::try_from(file_size).map_err(|_| Fat32Error::OutOfMemory)?;
-        let mut buf: Vec<u8> = alloc::vec![0u8; buf_size];
-        let mut total_read: usize = 0;
-
-        while total_read < buf.len() {
-            let n: usize = self.inner.read(&mut buf[total_read..])?;
-            if n == 0 {
-                break;
-            }
-            total_read += n;
-        }
-
-        buf.truncate(total_read);
-        Ok(buf)
-    }
-}
-
-impl core::fmt::Debug for File {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        f.debug_struct("File")
-            .field("mount_path", &self.mount_path)
-            .field("writable", &self.is_writable())
-            .finish_non_exhaustive()
-    }
-}
-
-impl Drop for File {
-    fn drop(&mut self) {
-        state::decrement_open_count(&self.mount_path);
-    }
-}
 
 //==================================================================================================
 // Public API Functions
@@ -403,39 +91,6 @@ pub fn file_raw_region(path: &str) -> Option<(*const u8, usize)> {
     .ok()
 }
 
-/// File metadata.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Stat {
-    /// Size of the file in bytes (0 for directories).
-    size: u64,
-    /// Whether this is a directory.
-    is_dir: bool,
-}
-
-impl Stat {
-    /// Creates a new `Stat` instance.
-    ///
-    /// # Parameters
-    ///
-    /// - `size`: File size in bytes (0 for directories).
-    /// - `is_dir`: Whether this entry is a directory.
-    pub fn new(size: u64, is_dir: bool) -> Self {
-        Self { size, is_dir }
-    }
-
-    /// Returns the file size in bytes (0 for directories).
-    #[must_use]
-    pub fn size(&self) -> u64 {
-        self.size
-    }
-
-    /// Returns whether this entry is a directory.
-    #[must_use]
-    pub fn is_dir(&self) -> bool {
-        self.is_dir
-    }
-}
-
 /// Gets file metadata without opening the file.
 ///
 /// # Parameters
@@ -463,48 +118,6 @@ pub fn stat(path: &str) -> Result<Stat, Fat32Error> {
         let fat_stat = mount.fat().stat(&relative_path)?;
         Ok(Stat::new(fat_stat.size, fat_stat.is_dir))
     })
-}
-
-/// Directory entry returned by [`read_dir()`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DirEntry {
-    /// Name of the entry (just the filename, not full path).
-    name: String,
-    /// Whether this entry is a directory.
-    is_dir: bool,
-    /// Size in bytes (0 for directories).
-    size: u64,
-}
-
-impl DirEntry {
-    /// Creates a new `DirEntry` instance.
-    ///
-    /// # Parameters
-    ///
-    /// - `name`: Entry name (filename only, not full path).
-    /// - `is_dir`: Whether this entry is a directory.
-    /// - `size`: Size in bytes (0 for directories).
-    pub fn new(name: String, is_dir: bool, size: u64) -> Self {
-        Self { name, is_dir, size }
-    }
-
-    /// Returns the entry name (filename only, not full path).
-    #[must_use]
-    pub fn name(&self) -> &str {
-        &self.name
-    }
-
-    /// Returns whether this entry is a directory.
-    #[must_use]
-    pub fn is_dir(&self) -> bool {
-        self.is_dir
-    }
-
-    /// Returns the size in bytes (0 for directories).
-    #[must_use]
-    pub fn size(&self) -> u64 {
-        self.size
-    }
 }
 
 /// Creates a directory.
@@ -832,141 +445,5 @@ fn open_with_options(
     // Track the open file (outside VFS lock to avoid nested locking).
     state::increment_open_count(&mount_path);
 
-    Ok(File {
-        inner: fat_file,
-        mount_path,
-    })
-}
-
-//==================================================================================================
-// Unit Tests
-//==================================================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // -- OpenOptions builder tests -----------------------------------------------
-
-    /// Tests that default OpenOptions has all flags false.
-    #[test]
-    fn open_options_default() {
-        let opts: OpenOptions = OpenOptions::new();
-        assert!(!opts.read, "read should default to false");
-        assert!(!opts.write, "write should default to false");
-        assert!(!opts.create, "create should default to false");
-        assert!(!opts.create_new, "create_new should default to false");
-        assert!(!opts.truncate, "truncate should default to false");
-    }
-
-    /// Tests builder method chaining.
-    #[test]
-    fn open_options_builder_chaining() {
-        let opts: OpenOptions = OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .truncate(true);
-        assert!(opts.read, "read should be true");
-        assert!(opts.write, "write should be true");
-        assert!(opts.create, "create should be true");
-        assert!(opts.truncate, "truncate should be true");
-    }
-
-    /// Tests that truncate without write is rejected.
-    #[test]
-    fn open_options_truncate_requires_write() {
-        let result = OpenOptions::new().truncate(true).open("/nonexistent");
-        assert_eq!(
-            result.unwrap_err(),
-            Fat32Error::InvalidArgument,
-            "truncate without write should be rejected"
-        );
-    }
-
-    /// Tests that create_new + create is rejected.
-    #[test]
-    fn open_options_create_new_excludes_create() {
-        let result = OpenOptions::new()
-            .write(true)
-            .create(true)
-            .create_new(true)
-            .open("/nonexistent");
-        assert_eq!(
-            result.unwrap_err(),
-            Fat32Error::InvalidArgument,
-            "create_new + create should be rejected"
-        );
-    }
-
-    /// Tests that create_new + truncate is rejected.
-    #[test]
-    fn open_options_create_new_excludes_truncate() {
-        let result = OpenOptions::new()
-            .write(true)
-            .truncate(true)
-            .create_new(true)
-            .open("/nonexistent");
-        assert_eq!(
-            result.unwrap_err(),
-            Fat32Error::InvalidArgument,
-            "create_new + truncate should be rejected"
-        );
-    }
-
-    /// Tests that Default trait matches OpenOptions::new().
-    #[test]
-    fn open_options_implements_default() {
-        let default: OpenOptions = OpenOptions::default();
-        let new: OpenOptions = OpenOptions::new();
-        assert_eq!(default.read, new.read);
-        assert_eq!(default.write, new.write);
-        assert_eq!(default.create, new.create);
-        assert_eq!(default.create_new, new.create_new);
-        assert_eq!(default.truncate, new.truncate);
-    }
-
-    /// Tests that OpenOptions implements Debug.
-    #[test]
-    fn open_options_debug() {
-        let opts: OpenOptions = OpenOptions::new().read(true);
-        let debug: alloc::string::String = alloc::format!("{opts:?}");
-        assert!(debug.contains("OpenOptions"), "debug output should contain type name");
-    }
-
-    // -- Stat tests --------------------------------------------------------------
-
-    /// Tests Stat equality and debug.
-    #[test]
-    fn stat_clone_eq_debug() {
-        let s: Stat = Stat::new(42, false);
-        let cloned: Stat = s;
-        assert_eq!(s, cloned, "clone should preserve equality");
-
-        let other: Stat = Stat::new(0, true);
-        assert_ne!(s, other, "different stats should not be equal");
-
-        assert_eq!(s.size(), 42, "size accessor should return 42");
-        assert!(!s.is_dir(), "is_dir accessor should return false");
-
-        let debug: alloc::string::String = alloc::format!("{s:?}");
-        assert!(debug.contains("42"), "debug should contain size");
-    }
-
-    // -- DirEntry tests ----------------------------------------------------------
-
-    /// Tests DirEntry equality and debug.
-    #[test]
-    fn dir_entry_clone_eq_debug() {
-        let entry: DirEntry = DirEntry::new(String::from("test.txt"), false, 100);
-        let cloned: DirEntry = entry.clone();
-        assert_eq!(entry, cloned, "clone should preserve equality");
-
-        assert_eq!(entry.name(), "test.txt", "name accessor should return name");
-        assert!(!entry.is_dir(), "is_dir accessor should return false");
-        assert_eq!(entry.size(), 100, "size accessor should return 100");
-
-        let debug: alloc::string::String = alloc::format!("{entry:?}");
-        assert!(debug.contains("test.txt"), "debug should contain name");
-    }
+    Ok(File::new(fat_file, mount_path))
 }
